requirementsLinux.txt
pandas==2.2.3
numpy==2.2.4
matplotlib==3.10.1
openpyxl==3.1.5
python-dotenv==1.1.0
langchain==0.3.23
langchain-groq==0.3.2
pymongo==4.12.1
langchain-core==0.3.51
plotly==6.0.1
scikit-learn==1.6.1
optuna==4.3.0
requests==2.32.3
nltk==3.9.1
sentence-transformers==4.0.2
